refactor: extract getCasesForDelta into data-display-utils
Move the delta hit-test helper (and its RTree / CaseObject types) out of
background.tsx into data-display-utils.ts, next to the rectangleSubtract it
calls, so the residual-plot marquee can reuse it. No behavior change.
